Arreglado caso de prueba 4

+++ src/main/java/sge/Coordenates.java
@@ -5,22 +5,22 @@ import javax.persistence.Embeddable;
 @Embeddable
 public class Coordenates {
 
-	float latitud;
-	float longitud;
+	Double latitud;
+	Double longitud;
 	
 	public Coordenates() {
 		super();
 	}
 	
-	public Coordenates(float longitud, float latitud) {
+	public Coordenates(Double longitud, Double latitud) {
 		this.latitud = latitud;
 		this.longitud = longitud;
 	}
-	public float X() {
+	public Double X() {
 		return longitud;
 	}
 	
-	public float Y() {
+	public Double Y() {
 		return latitud;
 	}
